fix(kernel): Wave 1 — eliminate one-step kernel kill/panic class (11 items)
Third-round review Wave 1 (docs/development/fix-plan-2026-09-12.md):

- ARCH-C1: uaccess.S exception table entries were bound one instruction
  past their fault sites (label-after-macro); every copy_from_user/
  clear_user fault escaped the table and panicked the kernel. Rebind all
  26 entries (label before insn, EXTABLE after); add
  scripts/verify_exception_table.py which cross-checks .ex_table against
  disassembly (26 entries, 0 errors); align .ex_table to 8 in linker.ld.
- ARCH-C2: trap_entry clobbered t0/t1/t2 before saving them (stack
  selection ran first). Rework entry Linux-style: dispatch via beqz/bltz
  (no temps), kernel path saves full GPR set on current stack (no switch,
  fixes nested-trap frame overwrite too), user path stashes user sp in
  ti_user_sp before switching. Binary-verified; SMP boot + interactive
  shell test pass.
- SYSA-C1/IOU-H3: clamp user-controlled read/write/getdents/io_uring
  lengths (MAX_RW_COUNT / 4MB) — huge counts no longer panic the
  allocator.
- PROC-P01/VFS-H4/H5/INIT-M10: validate ELF p_offset/p_filesz, segment
  vaddr range, and interpreter e_phoff (get_program_header central bounds
  check) — malformed ELFs return ENOEXEC instead of panicking.
- syscallb-C01: riscv_hwprobe no longer reads M-mode CSRs from S-mode;
  kernel-mode illegal instruction/ebreak/unknown exception now panic
  instead of silently skipping epc; print_cpu_info/get_core_id use the
  tp protocol.
- EXT4-H2/C1: mount validates on-disk geometry (block size, group/inode
  counts, desc size, total blocks vs device capacity) — no more
  division-by-zero panics or heap exhaustion on forged superblocks;
  extent tree bounds + depth<=5 (blocks forged eh_max/eh_entries OOB).
- MM-C1: copy_page_contents goes through phys_to_virt linear mapping.
- syscallb-H11/IOU-C1: epoll/timerfd/io_uring_enter/register verify file
  ops before casting private_data (type-confusion guard).
- TEST-H6: CI workflows set -o pipefail; spin Makefile exits non-zero on
  FAIL/COMPILE ERROR.
- 1.6 (partial): futex, syslog (kernel-buffer staging, no user writes
  under spinlock), setitimer, times, bind/connect, sigprocmask,
  sigaction now use exception-table copy_from/to_user paths.
- Fix pre-existing unit-test feature build breakage (13 type errors in
  tests); note: unit-test runtime fault confirmed pre-existing (stash
  A/B test) — tracked under Wave 8.

Verified: make build + release build, ex_table check 26/0, QEMU 4-CPU
SMP boot to mrsh shell, echo/ls interactive pass.

+++ kernel/src/tests/syscall_process.rs
@@ -295,7 +295,7 @@ fn test_sys_ids() {
         let expected_eperm = -(errno::EPERM as i64);
         test_assert_eq!(ret_setgid, expected_eperm,
             "sys_setgid(0) returns -EPERM without root");
-        let _ = sys_setgid([gid_before, 0, 0, 0, 0, 0]);
+        let _ = sys_setgid([gid_before as u64, 0, 0, 0, 0, 0]);
     }
 
     // getgroups(0, NULL) returns number of supplementary groups (0)
